move javascript extraction rules behind the indexer language registry
S3-JAVASCRIPT (ARCH-012), a one-language rollback unit following S3-KOTLIN.
JavaScript's tree-sitter rule file, compiled-rule cache, member-call callsite
marker, and receiver-call resolution engine move out of the indexer god file
into `codestory-indexer/src/languages/javascript.rs` as one
`LanguageExtraction` row. Every language-keyed dispatch that used to spell
"javascript" now reads the registry first and keeps its residual `match` for
the languages that have not moved; the last sibling package deletes those arms.

The move is mechanical. Thirteen functions left `lib.rs` and thirteen landed
in `languages/javascript.rs`; sorting both sides and diffing with blank lines
dropped leaves only the two declared renames
(`collect_javascript_receiver_call_edges` -> `receiver_call_specs`,
`javascript_member_call` -> `member_call`). The shared JS-family helpers stay
in `lib.rs` because TypeScript and TSX call them:
`js_like_callable_source_name`, `js_ts_visible_local_type_name`,
`js_ts_local_binding_visible_at_call`, `js_ts_lexical_scope`,
`normalize_js_ts_private_receiver_surface`,
`collect_javascript_static_call_edges`, and
`collect_javascript_runtime_import_specs`.

Proof: `tests/language_extraction_snapshot.rs` gains a `javascript` row plus
two goldens captured *before* the production edit and re-verified byte-identical
after (sha256 46324f56.. / 8444a609..). The threshold suites are not sufficient
on their own: changing JavaScript's qualified-name delimiter leaves both
`fidelity_regression` and `tictactoe_language_coverage` green and turns only the
new snapshots red.

The CLI comment roster and the indexer route-comment roster still disagree by
design (the CLI carries cpp/swift/tsx/jsx and not c; the route scanner carries
vue/astro and not cpp/swift/tsx). Only JavaScript's own entry moved out of each;
`jsx` stays in the CLI roster because it is a highlighter dialect with no
registry profile.

Declared behavior changes:
- None to any indexing, resolution, or rendering output. JavaScript
  fidelity_regression and tictactoe_language_coverage projections are proven
  byte-identical before and after, and packet_search_eval is unmoved.
- `LanguageRuleset::JavaScript`'s arm in `compiled_rules` is now unreachable and
  returns a typed error instead of compiled rules; the registry answers first.
  Fail-closed by construction, matching the Kotlin arm.
- New public row in
  `codestory_contracts::language_support::LANGUAGE_COMMENT_PROFILES` for
  `javascript`. Purely additive; `LANGUAGE_SUPPORT_PROFILES` is unchanged, so
  extension routing, claim tiers, and every existing language claim are
  identical.
- No version surface touched; everything stays 0.16.3.
